Bug 1129161 - Re-implement connecting to Firefox Account in Settings sheet.
This implements a simpler view controller which displays an interstitial
spinner while loading.  On success; the underlying webview fades in.  On
error, a simple error text fades in.  There's no affordance to reload;
UX can paint that bikeshed as they see fit.

This patch also reflects the state of the Firefox Account in the
settings panel.  The state doesn't change yet, but it's there.

https://github.com/mozilla/firefox-ios/commit/fb107a0515c492c7427e545fc9724946df84f7de