Bug 1164236 - Pre: provide localizable a11y strings
Because of the imminent string freeze, let's have the localizable
strings that will be probably required for full solution of the bug
available before the freeze. Then, when implementing a11y fixes
themselves later (after the string freeze), we will be able to just use
the already present localizable strings.

https://github.com/mozilla/firefox-ios/commit/e1b54c6d69b5d697037ecc8ed52ce9e01c732fc5