Bug 1315115 - Add toast to "Switch to tab" to the "Open in Tab" dialog. (#2437)
* Bug 1315115 - Add toast to "Switch to tab" to "Open in Tab". r?farhan

Open questions:
* Is there a danger keeping the reference to the Tab like this?  Can
  it disappear beneath me?
* Is the ButtonToast display logic correct?  Is this registering the
  toast appropriately?
* Are the strings localized correctly?
* How hard is it to add tests for this functionality?

* Review comments: Don't show toast when top tabs are visible; extract string constants.

I didn't extract additional strings that were nearby, since it looks
like this file isn't really using the Strings.* style yet.

I elected to duplicate the "Switch" button label since it may be that
some locales want to use different verbs for regular and private tabs;
or we may want to color or border the button differently to emphasize
that the new tab is private in future.  In that case, it'll be harder
if we try to unify the button label and styling now.

* refactor for swift 3.

https://github.com/mozilla/firefox-ios/commit/b7241b451f0da9d862f3b7c043f4e4de8c607be6